requirements.txt
bottle 
werkzeug 
python-socketio
eventlet